+++ frontend/src/Components/auth/Otp.css
@@ -1,4 +1,4 @@
-.wapper_otp {
+.wrapper_otp {
     margin: 0px;
     border: none;
     display: flex;
