+++ frontend/src/Components/auth/Signin.css
@@ -1,3 +1,19 @@
+.left_signup {
+    background: #40339f;
+    color: White;
+    border: 2px solid #40339f;
+    width: 35vw;
+    height: 100vh;
+    text-align: center;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    p {
+      color: #ffffffca;
+    }
+}
+
 .container_right.sign-up {
     /* height: 100vh; */
     overflow-y: auto;
